backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paddlepaddle==2.6.2
paddleocr==2.7.3
celery==5.3.4
redis==5.0.1
python-docx==1.1.0
openpyxl==3.1.2
pymupdf==1.23.8
opencv-python-headless==4.8.1.78
pillow==10.2.0
numpy==1.24.3
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==8.0.0
pytest-asyncio==0.23.3
beautifulsoup4==4.12.3
